+++ tests/python/test_vtu.py
@@ -5,7 +5,6 @@ import numpy as np
 import pytest
 
 import meshioplusplus
-
 from meshioplusplus.vtu import _vtu
 
 from . import helpers